\name{fBonds-package}


\alias{fBonds-package}
\alias{fBonds}


\docType{package}


\title{Pricing and Evaluating Bonds}


\description{

    The Rmetrics "fBonds" package is a collection of functions for 
    pricing and evaluating bonds and to compute term structures.
    
}


\details{

    \preformatted{
        Package:    \tab fBonds\cr
        Type:       \tab Package\cr
        Version:    \tab R 3.0.1\cr
        Date:       \tab 2014\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2014 Rmetrics Assiciation\cr
        URL:        \tab \url{https://www.rmetrics.org}
    }
    
}


\section{1 Introduction}{

    This section provides functions to valuate Bonds and to
    calculate term structures.
    
}


\section{2 Term Structure Calculation}{

    This section provides two functions for term structure modelling
    based on the Nelson-Siegel, and on the Svennson approach.
    
    \preformatted{
    NelsonSiegel    models Nelson-Siegel Term Structure
    Svensson        models Nelson-Siegel-Svensson Term Structure
    }
    
}


\section{About Rmetrics}{

    The \code{fBonds} Rmetrics package is written for educational 
    support in teaching "Computational Finance and Financial Engineering" 
    and licensed under the GPL.
    
}


\keyword{package}


